Moved batch integration test to a separated folder
